development continues

+++ bup/github_backup.py
@@ -8,7 +8,7 @@ from git import Repo
 from git.exc import GitCommandError, InvalidGitRepositoryError
 from balsa import get_logger
 
-from bup import __application_name__, BupBase, BackupTypes, get_preferences
+from bup import __application_name__, BupBase, BackupTypes, get_preferences, ExclusionPreferences
 
 log = get_logger(__application_name__)
 
@@ -23,8 +23,8 @@ def pull_branches(repo_name: str, branches: Iterable, repo_dir: str, log_functio
 
     if git_repo is not None:
         for branch in branches:
-            branch_name = branch.application_name
-            log_function(f'git pull "{repo_name}" branch:"{branch_name}" to {repo_dir}')
+            branch_name = branch.name
+            log_function(f'git pull "{repo_name}" branch:"{branch_name}"')
             git_repo.git.checkout(branch_name)
             git_repo.git.pull()
 
@@ -36,34 +36,39 @@ class GithubBackup(BupBase):
     def run(self):
 
         preferences = get_preferences()
+        excludes = ExclusionPreferences(BackupTypes.github.name).get()
 
         backup_dir = Path(preferences.backup_directory, "github")
         gh = github3.login(token=preferences.github_token)
         for github_repo in gh.repositories():
 
-            repo_name = str(github_repo)
-            repo_dir = Path(backup_dir, repo_name).absolute()
-            branches = github_repo.branches()
+            repo_owner_and_name = str(github_repo)
+            repo_name = repo_owner_and_name.split("/")[-1]
+            if any([e == repo_name for e in excludes]):
+                self.info_out(f"{repo_owner_and_name} excluded")
+            else:
+                repo_dir = Path(backup_dir, repo_owner_and_name).absolute()
+                branches = github_repo.branches()
 
-            # if we've cloned previously, just do a pull
-            pull_success = False
-            if repo_dir.exists():
-                try:
-                    pull_branches(repo_name, branches, repo_dir, self.info_out)
-                    pull_success = True
-                except GitCommandError as e:
-                    log.info(e)
-                    self.warning_out(f'could not pull "{repo_dir}" - will try to start over and do a clone of "{repo_name}"')
+                # if we've cloned previously, just do a pull
+                pull_success = False
+                if repo_dir.exists():
+                    try:
+                        pull_branches(repo_owner_and_name, branches, repo_dir, self.info_out)
+                        pull_success = True
+                    except GitCommandError as e:
+                        log.info(e)
+                        self.warning_out(f'could not pull "{repo_dir}" - will try to start over and do a clone of "{repo_owner_and_name}"')
 
-            # new to us - clone the repo
-            if not pull_success:
-                try:
-                    if repo_dir.exists():
-                        shutil.rmtree(repo_dir)
+                # new to us - clone the repo
+                if not pull_success:
+                    try:
+                        if repo_dir.exists():
+                            shutil.rmtree(repo_dir)
 
-                    self.info_out(f'git clone "{repo_name}" to "{repo_dir}"')
+                        self.info_out(f'git clone "{repo_owner_and_name}"')
 
-                    Repo.clone_from(github_repo.clone_url, repo_dir)
-                    pull_branches(repo_name, branches, repo_dir, self.info_out)
-                except PermissionError as e:
-                    log.warning(f"{repo_name} : {e}")
+                        Repo.clone_from(github_repo.clone_url, repo_dir)
+                        pull_branches(repo_owner_and_name, branches, repo_dir, self.info_out)
+                    except PermissionError as e:
+                        log.warning(f"{repo_owner_and_name} : {e}")
